requirements.txt
paho-mqtt
customtkinter
python-vlc
